Add signing key caching and sigv4 public proc docs
Add check for key expiration in AwsClient.request and create overloaded
versions of create_aws_authorization.  One which takes a key/id and
returns an AWS Authorization header string and another which takes
AwsCredentials and returns a Signing Key and adds an Authorization
string to the headers Table.

Provide the create_signing_key and a create_aws_authorization procs for
development and more advanced patterns with key caching.  Provide a
second create_aws_authorization using op overloading which will always
create a key, sign a request and add the authorization to the headers.